+++ test/reader_test.cpp
@@ -17,7 +17,7 @@ TEST_CASE("Reader tests", "[Reader]")
         SECTION("GetCopc Test")
         {
             auto copc = reader.GetCopcHeader();
-            REQUIRE(copc.span == 0);
+            REQUIRE(copc.span == 128);
             REQUIRE(copc.root_hier_offset == 93169718);
             REQUIRE(copc.root_hier_size == 8896);
             REQUIRE(copc.laz_vlr_offset == 643);
@@ -56,7 +56,7 @@ TEST_CASE("Reader tests", "[Reader]")
         SECTION("GetCopc Test")
         {
             auto copc = reader.GetCopcHeader();
-            REQUIRE(copc.span == 0);
+            REQUIRE(copc.span == 128);
             REQUIRE(copc.root_hier_offset == 93169718);
             REQUIRE(copc.root_hier_size == 8896);
             REQUIRE(copc.laz_vlr_offset == 643);
@@ -193,6 +193,10 @@ TEST_CASE("Spatial Query Functions", "[Reader]")
 {
     FileReader reader("autzen-classified.copc.laz");
 
+    // Make horizontal 2D box of [200,200] roughly in the middle of the point cloud.
+    auto middle = (reader.GetLasHeader().max + reader.GetLasHeader().min) / 2;
+    Box middle_box(middle.x - 200, middle.y - 200, middle.x + 200, middle.y + 200);
+
     SECTION("GetNodesWithinBox")
     {
         // Check that no nodes fit in a zero-sized box
@@ -207,8 +211,7 @@ TEST_CASE("Spatial Query Functions", "[Reader]")
 
     SECTION("GetNodesIntersectBox")
     {
-        // Take horizontal 2D box of [200,200] roughly in the middle of the point cloud.
-        auto subset_nodes = reader.GetNodesIntersectBox(Box(637190, 851109, 637390, 851309));
+        auto subset_nodes = reader.GetNodesIntersectBox(middle_box);
         REQUIRE(subset_nodes.size() == 13);
     }
 
@@ -233,8 +236,31 @@ TEST_CASE("Spatial Query Functions", "[Reader]")
         }
         {
             // Take horizontal 2D box of [200,200] roughly in the middle of the point cloud.
-            auto subset_points = reader.GetPointsWithinBox(Box(637190, 851109, 637390, 851309));
-            REQUIRE(subset_points.Get().size() == 22902);
+            auto subset_points = reader.GetPointsWithinBox(middle_box);
+            REQUIRE(subset_points.Get().size() == 91178);
         }
+    }
+
+    SECTION("GetDepthAtResolution")
+    {
+        REQUIRE(reader.GetDepthAtResolution(2) == 4);
+        REQUIRE(reader.GetDepthAtResolution(0) == 5);
+        REQUIRE(reader.GetDepthAtResolution(std::numeric_limits<double>::min()) == 5);
+        REQUIRE(reader.GetDepthAtResolution(std::numeric_limits<double>::max()) == 0);
+    }
+
+    SECTION("GetNodesAtResolution")
+    {
+        auto subset_nodes = reader.GetNodesAtResolution(2);
+        REQUIRE(subset_nodes.size() == 192);
+        for (const auto &node : reader.GetNodesAtResolution(0))
+            REQUIRE(node.key.d == 5);
+    }
+
+    SECTION("GetNodesWithinResolution")
+    {
+        auto subset_nodes = reader.GetNodesWithinResolution(2);
+        REQUIRE(subset_nodes.size() == 257);
+        REQUIRE(reader.GetNodesWithinResolution(0).size() == reader.GetAllChildren().size());
     }
 }