+++ test/vector3_test.cpp
@@ -32,7 +32,37 @@ TEST_CASE("Vector3", "[Vector3]")
         REQUIRE(vec.y == 0.0);
         REQUIRE(vec.z == 0.0);
     }
-    SECTION("Operators")
+    SECTION("Vector3 Operators")
+    {
+        auto vec = Vector3(1.0, 2.0, 3.0);
+        auto other = Vector3(1.0, 2.0, 3.0);
+
+        auto res = vec * other;
+
+        REQUIRE(res.x == 1.0);
+        REQUIRE(res.y == 4.0);
+        REQUIRE(res.z == 9.0);
+
+        res = vec / other;
+
+        REQUIRE(res.x == 1.0);
+        REQUIRE(res.y == 1.0);
+        REQUIRE(res.z == 1.0);
+
+        res = vec + other;
+
+        REQUIRE(res.x == 2.0);
+        REQUIRE(res.y == 4.0);
+        REQUIRE(res.z == 6.0);
+
+        res = vec - other;
+
+        REQUIRE(res.x == 0.0);
+        REQUIRE(res.y == 0.0);
+        REQUIRE(res.z == 0.0);
+    }
+
+    SECTION("Double Operators")
     {
         auto vec = Vector3(1.0, 2.0, 3.0);
         auto vec2 = vec;